refactor endpoint v2 trace labels
